pyproject.toml
[project]
name = "smart-video-trimmer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-whisper>=20250625",
    "webrtcvad>=2.0.10",
]
